+++ src/app/(route)/layout.tsx
@@ -1,0 +1,10 @@
+import { Navigation } from '@/shared/components';
+
+export default function RouteLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <>
+      {children}
+      <Navigation />
+    </>
+  );
+}
