+++ src/app/(route)/link/page.tsx
@@ -1,0 +1,3 @@
+export default function LinkPage() {
+  return <span>link</span>;
+}
