+++ src/app/(route)/memo/components/MemoItem.tsx
