+++ src/app/(route)/memo/components/MemoPreview.tsx
@@ -1,4 +1,4 @@
-import { Memo } from '../../api/memos/[memoListId]/interface';
+import { Memo } from '../../../api/memos/[memoListId]/interface';
 import { cn } from '@/shared/utils/cn';
 import dayjs from 'dayjs';
 
@@ -12,11 +12,11 @@ const MemoPreview = ({ title, createdAt, tags, className, onClick }: MemoPreview
 
   return (
     <div
-      className={cn('w-100 border-grayscale-200 flex flex-col gap-2.5 border-solid bg-white p-3.5', className)}
+      className={cn('w-100 flex flex-col gap-2.5 border-solid border-grayscale-200 bg-white p-3.5', className)}
       onClick={onClick}
     >
       <h4 className='truncate font-bold'>{title}</h4>
-      <p className='text-grayscale-500 text-xs'>{formattedDate}</p>
+      <p className='text-xs text-grayscale-500'>{formattedDate}</p>
       <div className='flex flex-nowrap gap-1'>
         {tags.length > 0 && (
           <>
@@ -24,7 +24,7 @@ const MemoPreview = ({ title, createdAt, tags, className, onClick }: MemoPreview
               {`#${tags[0].name}`}
             </div>
             {tags.length > 1 && (
-              <div className='bg-grayscale-100 shrink-0 grow-0 basis-auto rounded-xl px-2 py-1 text-[0.625rem]'>
+              <div className='shrink-0 grow-0 basis-auto rounded-xl bg-grayscale-100 px-2 py-1 text-[0.625rem]'>
                 {`+${tags.length - 1}`}
               </div>
             )}
