+++ src/app/(route)/todo/components/TodoItem.tsx
@@ -3,8 +3,8 @@
 import { useSearchParams } from 'next/navigation';
 import { type FocusEvent, type FormEvent, useCallback, useRef, useState } from 'react';
 import { TodoContentEditableText } from './TodoContentEditableText';
-import { useDeleteTodo } from '@/app/todo/queries/useDeleteTodo';
-import { useUpdateTodo } from '@/app/todo/queries/useUpdateTodo';
+import { useDeleteTodo } from '@/app/(route)/todo/queries/useDeleteTodo';
+import { useUpdateTodo } from '@/app/(route)/todo/queries/useUpdateTodo';
 import { Button, CheckBox, Modal } from '@/shared/components';
 
 interface TodoItemProps {
@@ -67,23 +67,21 @@ export function TodoItem({ id, isFullfilled, content, createdAt }: TodoItemProps
     <form className='flex w-full items-start gap-2 py-3' onSubmit={handleSubmit}>
       <CheckBox isChecked={isFullfilled} onClick={() => toggleCheck()} className='shrink-0' />
       <TodoContentEditableText textRef={textRef} onFocus={() => setIsEditMode(true)} onBlur={handleBlurText} />
-      <button ref={buttonRef} type='submit' className='text-body1 text-grayscale-700 w-fit shrink-0'>
+      <button ref={buttonRef} type='submit' className='w-fit shrink-0 text-body1 text-grayscale-700'>
         {isEditMode ? '확인' : '삭제'}
       </button>
-      {
-        <Modal
-          isOpen={isModalOpen}
-          title='삭제하실건가요?'
-          message='삭제한 내용은 되돌릴 수 없어요'
-          firstButton={<Button onClick={() => deleteTodo(id)}>확인</Button>}
-          secondButton={
-            <Button color='secondary' onClick={() => closeModal()}>
-              취소
-            </Button>
-          }
-          close={() => closeModal()}
-        />
-      }
+      <Modal
+        isOpen={isModalOpen}
+        title='삭제하실건가요?'
+        message='삭제한 내용은 되돌릴 수 없어요'
+        firstButton={<Button onClick={() => deleteTodo(id)}>확인</Button>}
+        secondButton={
+          <Button color='secondary' onClick={() => closeModal()}>
+            취소
+          </Button>
+        }
+        close={() => closeModal()}
+      />
     </form>
   );
 }
