+++ src/app/(route)/todo/queries/useCreateTodo.ts
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { todos } from './queryKey';
-import { createTodo } from '@/app/todo/apis/todo';
+import { createTodo } from '@/app/(route)/todo/apis/todo';
 
 export function useCreateTodo(when: string) {
   const queryClient = useQueryClient();
