+++ src/shared/components/Portal.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import type { PropsWithChildren } from 'react';
 import { useRef } from 'react';
 import { createPortal } from 'react-dom';
