+++ src/shared/components/index.ts
@@ -1,6 +1,6 @@
 export * from './Tabs';
 export * from './Icon';
-export { default as Navigation } from './Navigation';
+export * from './Navigation';
 export * from './Button';
 export * from './CheckBox';
 export * from './Portal';
